Command Line Dungeon Crawler
We implemented several different patterns to make a simplistic dungeon
crawler. One of the aspects of this project I was proudest of was the
maze and map. The maze is generated randomly every time you play and
the map is set up to show you visually what you’re looking at.